tests/wpt/mozilla/tests/css/float_relative_to_position.html

<!doctype html>
<html>
<head>
<meta charset=utf-8>
<title>Tests the relationship between float and position</title>
<head>
<body>
<div style="position: fixed; float: left"></div>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
    test(function() {
        var computed_style = getComputedStyle(document.querySelector("div"))["float"];
        assert_equals(computed_style, "none");

        document.querySelector("div").setAttribute("position", "absolute");

        computed_style = getComputedStyle(document.querySelector("div"))["float"];
        assert_equals(computed_style, "none");
        });

    <!-- test(); -->
</script>

</body>
</html>
